Move `RustFunctionGenerator` out of `rust-lib`
